--- source/dotnet/EDI.UnitTests/InboxEvents/Mappers/QuantityQualityMapperTests.cs
@@ -1,56 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using EDI.InboxEvents.Mappers;
-using Energinet.DataHub.Edi.Responses;
-using FluentAssertions;
-using Xunit;
-
-namespace Energinet.DataHub.Wholesale.EDI.UnitTests.InboxEvents.Mappers;
-
-public class QuantityQualityMapperTests
-{
-    [Theory]
-    [InlineData(Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.QuantityQuality.Calculated, QuantityQuality.Calculated)]
-    [InlineData(Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.QuantityQuality.Missing, QuantityQuality.Missing)]
-    [InlineData(Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.QuantityQuality.Measured, QuantityQuality.Measured)]
-    [InlineData(Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.QuantityQuality.Estimated, QuantityQuality.Estimated)]
-    [InlineData(Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.QuantityQuality.Incomplete, QuantityQuality.Incomplete)]
-    public void MapQuantityQuality_WhenCalled_MapsCorrectly(
-        Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.QuantityQuality quantityQuality,
-        QuantityQuality expected)
-    {
-        // Act & Assert
-        QuantityQualityMapper.MapQuantityQuality(quantityQuality).Should().Be(expected);
-    }
-
-    [Theory]
-    [MemberData(nameof(QuantityQualitys))]
-    public void MapQuantityQuality_WhenCalled_HandlesExpectedType(Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.QuantityQuality quantityQuality)
-    {
-        // Act
-        var actual = () => QuantityQualityMapper.MapQuantityQuality(quantityQuality);
-
-        // Assert
-        actual.Should().NotThrow();
-    }
-
-    public static IEnumerable<object[]> QuantityQualitys()
-    {
-        foreach (var quantityQuality in Enum.GetValues(typeof(Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.QuantityQuality)))
-        {
-            yield return new[] { quantityQuality };
-        }
-    }
-}
