+++ source/dotnet/EDI.UnitTests/TimeSeriesTypeMapperTest.cs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using EDI;
-using EDI.InboxEvents;
+using Energinet.DataHub.Wholesale.EDI.UnitTests.Mappers;
+using Energinet.DataHub.Wholesale.EDI.UnitTests.Models;
 using FluentAssertions;
 using Xunit;
 using CalculationTimeSeriesType = Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model.TimeSeriesType;
@@ -31,7 +31,7 @@ public class TimeSeriesTypeMapperTests
     public void ToCalculationTimeSerieType_ReturnsExpectedType(TimeSeriesType type, CalculationTimeSeriesType expected)
     {
         // Act
-        var actual = TimeSeriesTypeMapper.MapTimeSerieType(type);
+        var actual = TimeSeriesTypeMapper.MapTimeSeriesTypeFromEdi(type);
 
         // Assert
         actual.Should().Be(expected);
