+++ source/dotnet/wholesale-api/EDI/Client/EdiClient.cs
@@ -16,7 +16,7 @@ using Azure.Messaging.ServiceBus;
 using Energinet.DataHub.Wholesale.Events.Application.Options;
 using Microsoft.Extensions.Options;
 
-namespace EDI.InboxEvents;
+namespace Energinet.DataHub.Wholesale.EDI.UnitTests.Client;
 
 public class EdiClient : IEdiClient, IAsyncDisposable
 {
@@ -32,7 +32,9 @@ public class EdiClient : IEdiClient, IAsyncDisposable
         await _sender.SendMessageAsync(message, cancellationToken).ConfigureAwait(false);
     }
 
+#pragma warning disable VSTHRD200
     public async ValueTask DisposeAsync()
+#pragma warning restore VSTHRD200
     {
         await _sender.DisposeAsync().ConfigureAwait(false);
         GC.SuppressFinalize(this);
