+++ source/dotnet/wholesale-api/EDI/Contracts/AggregatedTimeSeriesRequest.proto
