+++ source/dotnet/wholesale-api/EDI/Contracts/AggregatedTimeSeriesRequestAccepted.proto
