+++ source/dotnet/wholesale-api/EDI/Contracts/AggregatedTimeSeriesRequestRejected.proto
