+++ source/dotnet/wholesale-api/EDI/Models/AggregatedTimeSeriesRequest.cs
@@ -12,15 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Azure.Messaging.ServiceBus;
+namespace Energinet.DataHub.Wholesale.EDI.UnitTests.Models;
 
-namespace EDI.InboxEvents;
-
-public interface IAggregatedTimeSeriesRequestFactory
-{
-    /// <summary>
-    /// Responsible for parsing the received message.
-    /// </summary>
-    /// <param name="request"></param>
-    AggregatedTimeSeriesRequest Parse(ServiceBusReceivedMessage request);
-}
+public record AggregatedTimeSeriesRequest(
+    Period Period,
+    TimeSeriesType TimeSeriesType,
+    AggregationPerRoleAndGridArea AggregationPerRoleAndGridArea);
