+++ source/dotnet/wholesale-api/EDI/Models/AggregationPerRoleAndGridArea.cs
@@ -12,8 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using NodaTime;
+namespace Energinet.DataHub.Wholesale.EDI.UnitTests.Models;
 
-namespace EDI.InboxEvents;
-
-public record Period(Instant Start, Instant End);
+public record AggregationPerRoleAndGridArea(string GridAreaCode, string? EnergySupplierId = null, string? BalanceResponsibleId = null);
