+++ source/dotnet/wholesale-api/Events/Events.IntegrationTests/Infrastructure/AggregatedTimeSeriesRequests/AggregatedTimeSeriesRequestsTests.cs
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 using Azure.Messaging.ServiceBus;
-using EDI;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
+using Energinet.DataHub.Wholesale.EDI.UnitTests;
 using Energinet.DataHub.Wholesale.Events.Application.UseCases;
 using Energinet.DataHub.Wholesale.Events.Application.Workers;
 using Energinet.DataHub.Wholesale.Events.IntegrationTests.Fixture;
